Inspect deviations: Attention
If you compare small numbers with only rtol-argument of np.allclose set, than you might get into trouble, as very small deviations of small numbers are "relatively" large.

If you set material properties randomly, you might inspect edge cases which are not that likely. Is this what you want?

What is the expected precision on components, which in the case of spherical inclusions are likely to be near zero, or even zero in the limit?